ApiController is now abstract, only providing publishAndCreateResponse
